requirements.txt
fastapi
mangum
uvicorn
pydantic
numpy
matplotlib
nltk
tensorflow
keras
pytest
scikit-learn
pandas
igraph
python-dotenv
python-multipart
pydantic[email]
authlib
werkzeug
flask
requests
python-jose
boto3
moto[all]
httpx